requirements.txt
asarPy==1.0.1
numpy==2.1.1
matplotlib==3.9.2
pillow==10.4.0
